vehicle_api/requirements.txt
flask
pandas
numpy
PyJWT
requests==2.31.0
pytest==7.4.3
gunicorn
gevent
python-dotenv
scikit-learn==1.3.2